refactor: ShoppingDetailScreen -> ShoppingDetailRoute 분리 및 state 전달로직 수정
