+++ app/src/main/java/nextstep/shoppingcart/ui/shoppingdetail/ShoppingDetailRoute.kt
@@ -1,0 +1,23 @@
+package nextstep.shoppingcart.ui.shoppingdetail
+
+import androidx.compose.runtime.Composable
+import nextstep.shoppingcart.data.Cart.addOne
+import nextstep.shoppingcart.data.Cart.findProductById
+
+@Composable
+fun ShoppingDetailRoute(
+    productId: Long,
+    onBackClick: () -> Unit,
+    onAddClick: () -> Unit,
+) {
+    val product = findProductById(productId)
+
+    ShoppingDetailScreen(
+        product = product,
+        onBackClick = onBackClick,
+        onAddClick = {
+            addOne(product)
+            onAddClick()
+        },
+    )
+}
